+++ src/interface.py
@@ -1,10 +1,19 @@
 import PySimpleGUI as psg
 import os
+import sys
 import pyperclip
-from osrparse import Replay, Mod
+import numpy as np
+import matplotlib.pyplot as plt
+from matplotlib.lines import Line2D
+from matplotlib.backends.backend_tkagg import FigureCanvasTkAgg
+from math import ceil
 from pyosudb import parse_osudb
+from copy import deepcopy
 from pathlib import Path
 from datetime import datetime
+from osrparse import Replay, Mod
+from osrparse.utils import LifeBarState
+from typing import Self
 
 import utils
 import calculation
@@ -12,100 +21,201 @@ from config import CONFIG
 
 
 def generate_mods_checkboxes(width: int):
+    mod_list = utils.mods_list()
+
+    height = ceil(len(mod_list) / width)
+
     layout = []
-    tmp_layout = []
-    x = 0
-    for mod in list(utils.MODS_2_CODES.keys())[1:]:
-        if x > width:
-            layout.append(tmp_layout)
-            tmp_layout = []
-            x = 0
+    tmp_text_column = []
+    tmp_checkbox_column = []
 
-        tmp_layout.append(psg.Text(mod))
-        tmp_layout.append(psg.Checkbox("", key=f"-MOD_{mod.upper()}-", default=False))
-        x += 1
-    if tmp_layout != []:
-        layout.append(tmp_layout)
-    return layout
+    mod_index = 0
+
+    for _ in range(width):
+        for _ in range(height):
+            try:
+                mod = mod_list[mod_index]
+                mod_index += 1
+            except IndexError:
+                mod_index += 1
+                continue
+
+            tmp_text_column.append(psg.Text(mod, pad=((0, 0), (3, 7)), justification="right", size=(10, 1)))
+            tmp_checkbox_column.append(psg.Checkbox("", key=f"-MOD_{mod.upper()}-"))
+
+        tmp_column = psg.Column([[psg.Column([[text] for text in tmp_text_column]), psg.Column([[checkbox] for checkbox in tmp_checkbox_column])]])
+        tmp_text_column = []
+        tmp_checkbox_column = []
+        layout.append(tmp_column)
+
+    return [layout]
 
 
-class App:
+class TemplateWindow:
+    LAYOUT = [[]]
+    EVENT_FUNCS = {}
+
+    WINDOW_NAME = "template"
+    WINDOW_SIZE = ()
+    WINDOW_TIMEOUT = 100
+
     def __init__(self) -> None:
-        self.window = self.create_window()
+        if len(self.WINDOW_SIZE) != 2:
+            self.window = psg.Window(self.WINDOW_NAME, deepcopy(self.LAYOUT), finalize=True)
+        else:
+            self.window = psg.Window(self.WINDOW_NAME, deepcopy(self.LAYOUT), finalize=True, size=self.WINDOW_SIZE)
+
+        self.EVENT_FUNCS = {}  # should be re-defined in __init__
+
+    def main_loop(self) -> None:
+        while True:
+            self._loop_step()
+
+    def _start(self) -> None:
+        self.main_loop()
+
+    def _loop_step(self) -> None:
+        self.event, self.values = self.window.read(timeout=self.WINDOW_TIMEOUT)
+
+        if self.event == psg.WIN_CLOSED or self.event == "-EXIT-":
+            sys.exit(0)
+
+        self.event = self.event.split("::")[-1:][0]  # bruh
+
+        tmp_func = self.EVENT_FUNCS.get(self.event)
+        if tmp_func is not None:
+            tmp_func()
+
+    def change_window(self, window: Self):
+        self.window.close()
+        del self
+        window._start()
+
+    @utils.run_async
+    def open_window(self, window: Self):
+        window._start()
+
+
+class LifeGraphEditor(TemplateWindow):
+    LAYOUT = [[psg.Canvas(key="-LIFE_GRAPH-", expand_x=True)]]
+
+    WINDOW_NAME = "Replay Editor: Life Graph"
+    WINDOW_SIZE = (1500, 400)
+    WINDOW_TIMEOUT = 10
+
+    def __init__(self, life_graph_data: list[LifeBarState]) -> None:
+        super().__init__()
+
+        self.life_graph_data = life_graph_data
+
+        self.fig, self.axes = plt.subplots()
+
+        self.line = Line2D([x.time for x in self.life_graph_data], [x.life * 100 for x in self.life_graph_data], marker="o", markeredgecolor="r", markersize=4)
+
+        self.axes.add_line(self.line)
+
+        # self.cid = self.poly.add_callback()
+
+        # self.canvas.mpl_connect('draw_event', self.on_draw)
+        # self.canvas.mpl_connect('button_press_event', self.on_button_press)
+        # self.canvas.mpl_connect('key_press_event', self.on_key_press)
+        # self.canvas.mpl_connect('button_release_event', self.on_button_release)
+        # self.canvas.mpl_connect('motion_notify_event', self.on_mouse_move)
+
+        # self.canvas = canvas
+        self.axes.set_xlim(0, self.life_graph_data[-1:][0].time)
+        self.axes.set_ylim(0, 100)
+        plt.xlabel('Ticks')
+        plt.ylabel('%')
+        plt.grid()
+
+        figure_canvas_agg = FigureCanvasTkAgg(self.fig, master=self.window['-LIFE_GRAPH-'].TKCanvas)
+        figure_canvas_agg.draw()
+        figure_canvas_agg.get_tk_widget().pack(side='right', fill='both', expand=1)
+
+    def _loop_step(self) -> None:
+        super()._loop_step()
+
+        self.update_info()
+
+    def update_info(self) -> None:
+        pass
+
+
+class ReplayEditor(TemplateWindow):
+    WINDOW_NAME = "Replay Editor: Replay data"
+
+
+class AttributesEditor(TemplateWindow):
+    MENU_LAYOUT = [['&File', ['&Open::-OPEN_REPLAY-', '&Save::-SAVE-', '&Save As...::-SAVEAS-', '---', 'E&xit::-EXIT-']],
+                   ['&Tools', ['&Life graph editor::-OPEN_LIFE_GRAPH_EDITOR-']]]
+
+    CHANGING_TEXT_COL = psg.Column([[psg.Text("Player", pad=((0, 0), (0, 65)))],
+                                    [psg.Text("300s")],
+                                    [psg.Text("100s")],
+                                    [psg.Text("50s")],
+                                    [psg.Text("Gekis")],
+                                    [psg.Text("Katus")],
+                                    [psg.Text("Misses")],
+                                    [psg.Text("Total score")],
+                                    [psg.Text("Max combo")],
+                                    [psg.Text("Perfect combo")],
+                                    [psg.Text("Date")]])
+
+    CHANGING_INPUT_COL = psg.Column([[psg.Multiline(default_text="", key="-USERNAME-", size=(50, 5)), psg.Text("", key="-USERNAME_ERROR-")],
+                                     [psg.Input(default_text="0", key="-N300-", size=(12, 1)), psg.Text("", key="-N300_ERROR-")],
+                                     [psg.Input(default_text="0", key="-N100-", size=(12, 1)), psg.Text("", key="-N100_ERROR-")],
+                                     [psg.Input(default_text="0", key="-N50-", size=(12, 1)), psg.Text("", key="-N50_ERROR-")],
+                                     [psg.Input(default_text="0", key="-NGEKIS-", size=(12, 1)), psg.Text("", key="-NGEKIS_ERROR-")],
+                                     [psg.Input(default_text="0", key="-NKATUS-", size=(12, 1)), psg.Text("", key="-NKATUS_ERROR-")],
+                                     [psg.Input(default_text="0", key="-NMISSES-", size=(12, 1)), psg.Text("", key="-NMISSES_ERROR-")],
+                                     [psg.Input(default_text="0", key="-TOTAL_SCORE-", size=(12, 1)), psg.Text("", key="-TOTAL_SCORE_ERROR-")],
+                                     [psg.Input(default_text="0", key="-MAX_COMBO-", size=(12, 1)), psg.Text("", key="-MAX_COMBO_ERROR-")],
+                                     [psg.Checkbox("", key="-PFC-",)],
+
+                                     [psg.CalendarButton("", key="-TIMESTAMP_DATE-", target=(psg.ThisRow, 0), close_when_date_chosen=False, format="%d/%m/%Y"),
+                                     psg.Spin(list(range(0, 24)), key="-TIMESTAMP_HOUR-", initial_value=datetime.today().strftime("%H")[:1], size=(2, 1)),
+                                     psg.Spin(list(range(0, 60)), key="-TIMESTAMP_MINUTE-", initial_value=datetime.today().strftime("%M"), size=(2, 1)),
+                                     psg.Spin(list(range(0, 60)), key="-TIMESTAMP_SECOND-", initial_value=datetime.today().strftime("%S"), size=(2, 1))]])
+
+    CHANGING_COL = psg.Column([[CHANGING_TEXT_COL, CHANGING_INPUT_COL],
+                               [psg.Frame("Mods", generate_mods_checkboxes(5))]])
+
+    INFO_COL = psg.Column([[psg.Text("Beatmap: "), psg.Text("None", key="-INFO_BEATMAP-")],
+                           [psg.Text("Total accuracy: "), psg.Text("None", key="-INFO_ACCURACY-")],
+                           [psg.Text("Total PP: "), psg.Text("None", key="-INFO_PP-")]],
+                          expand_y=True)
+
+    MAIN_COL = psg.Column([[CHANGING_COL, INFO_COL]])
+
+    LAYOUT = [[psg.Menu(MENU_LAYOUT)],
+              [MAIN_COL],
+              [psg.Text("CLI command")],
+              [psg.InputText(key="-CLI_COMMAND-", readonly=True, expand_x=True), psg.Button("Copy", key="-COPY_CLI-")]]
+
+    WINDOW_NAME = "Replay Editor: Attributes"
+
+    WINDOW_TIMEOUT = 10
+
+    def __init__(self) -> None:
+        super().__init__()
+
         self.osu_db = parse_osudb(Path(CONFIG.get("osu_path")) / "osu!.db")
         self.replay_path = None
         self.replay = None
 
-        self.BUTTON_FUNCS = {
+        self.EVENT_FUNCS = {
             "-OPEN_REPLAY-": lambda: self.open_replay(),
             "-SAVE-": lambda: self.save_replay(),
             "-SAVEAS-": lambda: self.save_as_replay(),
             "-COPY_CLI-": lambda: self.clipboard_copy_CLI_command(),
+            "-OPEN_LIFE_GRAPH_EDITOR-": lambda: self.open_life_graph_editor(),
             }
 
-    def main_loop(self) -> None:
-        while True:
-            self.event, self.values = self.window.read(timeout=10)
+    def _loop_step(self) -> None:
+        super()._loop_step()
 
-            if self.event == psg.WIN_CLOSED or self.event == "-EXIT-":
-                break
-
-            self.event = self.event.split("::")[-1:][0]  # bruh
-
-            try:
-                self.BUTTON_FUNCS.get(self.event)()
-            except TypeError:
-                pass
-
-            self.update_info()
-
-    def create_window(self):
-        MENU_LAYOUT = [['&File', ['&Open::-OPEN_REPLAY-', '&Save::-SAVE-', '&Save As...::-SAVEAS-', '---', 'E&xit::-EXIT-']]]
-
-        CHANGING_TEXT_COL = psg.Column([[psg.Text("Player", pad=((0, 0), (0, 65)))],
-                                        [psg.Text("300s")],
-                                        [psg.Text("100s")],
-                                        [psg.Text("50s")],
-                                        [psg.Text("Gekis")],
-                                        [psg.Text("Katus")],
-                                        [psg.Text("Misses")],
-                                        [psg.Text("Total score")],
-                                        [psg.Text("Max combo")],
-                                        [psg.Text("Perfect combo")],
-                                        [psg.Text("Date")]])
-
-        CHANGING_INPUT_COL = psg.Column([[psg.Multiline(default_text="", key="-USERNAME-", size=(50, 5)), psg.Text("", key="-USERNAME_ERROR-")],
-                                         [psg.Input(default_text="0", key="-N300-", size=(12, 1)), psg.Text("", key="-N300_ERROR-")],
-                                         [psg.Input(default_text="0", key="-N100-", size=(12, 1)), psg.Text("", key="-N100_ERROR-")],
-                                         [psg.Input(default_text="0", key="-N50-", size=(12, 1)), psg.Text("", key="-N50_ERROR-")],
-                                         [psg.Input(default_text="0", key="-NGEKIS-", size=(12, 1)), psg.Text("", key="-NGEKIS_ERROR-")],
-                                         [psg.Input(default_text="0", key="-NKATUS-", size=(12, 1)), psg.Text("", key="-NKATUS_ERROR-")],
-                                         [psg.Input(default_text="0", key="-NMISSES-", size=(12, 1)), psg.Text("", key="-NMISSES_ERROR-")],
-                                         [psg.Input(default_text="0", key="-TOTAL_SCORE-", size=(12, 1)), psg.Text("", key="-TOTAL_SCORE_ERROR-")],
-                                         [psg.Input(default_text="0", key="-MAX_COMBO-", size=(12, 1)), psg.Text("", key="-MAX_COMBO_ERROR-")],
-                                         [psg.Checkbox("", key="-PFC-",)],
-
-                                         [psg.CalendarButton("", key="-TIMESTAMP_DATE-", target=(psg.ThisRow, 0), close_when_date_chosen=False, format="%d/%m/%Y"),
-                                          psg.Spin(list(range(0, 24)), key="-TIMESTAMP_HOUR-", initial_value=datetime.today().strftime("%H")[:1], size=(2, 1)),
-                                          psg.Spin(list(range(0, 60)), key="-TIMESTAMP_MINUTE-", initial_value=datetime.today().strftime("%M"), size=(2, 1)),
-                                          psg.Spin(list(range(0, 60)), key="-TIMESTAMP_SECOND-", initial_value=datetime.today().strftime("%S"), size=(2, 1))]])
-
-        CHANGING_COL = psg.Column([[CHANGING_TEXT_COL, CHANGING_INPUT_COL],
-                                   [psg.Frame("Mods", generate_mods_checkboxes(6))]])
-
-        INFO_COL = psg.Column([[psg.Text("Beatmap: "), psg.Text("None", key="-INFO_BEATMAP-")],
-                               [psg.Text("Total accuracy: "), psg.Text("None", key="-INFO_ACCURACY-")],
-                               [psg.Text("Total PP: "), psg.Text("None", key="-INFO_PP-")]],
-                              expand_y=True)
-
-        MAIN_COL = psg.Column([[CHANGING_COL, INFO_COL]])
-
-        MAIN_LAYOUT = [[psg.Menu(MENU_LAYOUT)],
-                       [MAIN_COL],
-                       [psg.Text("CLI command")],
-                       [psg.InputText(key="-CLI_COMMAND-", readonly=True, expand_x=True), psg.Button("Copy", key="-COPY_CLI-")]]
-
-        return psg.Window("Osu replay editor", MAIN_LAYOUT, finalize=True, resizable=True)
+        self.update_info()
 
     def check_info(self):
         for value_name, value_of_value in self.values.items():
@@ -130,6 +240,7 @@ class App:
             beatmap = self.osu_db.get_beatmap_from_hash(self.replay.beatmap_hash)
             beatmap_path = str(Path(CONFIG.get("osu_path")) / "songs" / beatmap.folder_name / beatmap.osu_file)
             pp = calculation.calculate_pp(beatmap_path, mode=self.replay.mode, mods=self.get_mods(), n_geki=n_geki, n_katu=n_katu, n300=n300, n100=n100, n50=n50, n_misses=nmiss, combo=max_combo)
+
             self.window["-INFO_PP-"].update(f"{str(pp)}pp")
 
         self.window["-CLI_COMMAND-"].update(self.generate_CLI_command())
@@ -177,7 +288,7 @@ class App:
 
     def get_mods(self):
         mods = []
-        for mod in list(utils.MODS_2_CODES.keys())[1:]:
+        for mod in utils.mods_list():
             if self.values[f"-MOD_{mod.upper()}-"] is True:
                 mods.append(mod)
         return Mod(utils.mods2code(mods))
@@ -228,11 +339,18 @@ class App:
             beatmap = self.osu_db.get_beatmap_from_hash(self.replay.beatmap_hash)
             self.window["-INFO_BEATMAP-"].update(f"{beatmap.artist} // {beatmap.mapper} - {beatmap.title}")
 
-            for mod in list(utils.MODS_2_CODES.keys())[1:]:
+            for mod in utils.mods_list():
                 self.window[f"-MOD_{mod.upper()}-"].update(True if Mod(utils.mods2code([mod])) in self.replay.mods else False)
+
+    def open_life_graph_editor(self):
+        if self.replay is None:
+            self.show_error("Please, open replay first")
+            return
+
+        self.open_window(LifeGraphEditor(self.replay.life_bar_graph))
 
 
 if __name__ == "__main__":
-    app = App()
+    app = AttributesEditor()
 
-    app.main_loop()
+    app._start()